Update
replaced the bob with the  custom name .Converted pounds into stone and fahrenheit into centigrade. Replaced the 94 fahrenheit to the value  temperature  in centigrade and 300 pounds to the value of weight in stone so that when the user hit the different radio button it shows different value.
Lastly set story.textContent = newStory