Update
included an appropriate aria-label attribute on <form> input elements. Ensured the label text is wrapped inside <label> elements setting for attributes that contain the ID values of the input elements of the labels they are for. Made the show/hide comment control button keyboard accessible. Change the submit comment the button.